# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Order Book API
  version: "1.0.0"
  description: |
    Thread-safe order book with price-time priority matching engine.
    Supports Limit, Market, and IOC orders with self-match prevention,
    idempotency keys, and graceful shutdown.

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /healthz:
    get:
      summary: Health check
      operationId: healthz
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /orders:
    post:
      summary: Place an order
      operationId: placeOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
            format: uuid
          required: false
          description: |
            Client-generated idempotency key. When provided, duplicate
            requests within the TTL window (5 min) return the cached
            response of the first successful execution.
      responses:
        "200":
          description: Order accepted (may produce trades)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceOrderResponse"
        "400":
          description: Validation error, insufficient funds, or market not live
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    delete:
      summary: Cancel a resting order
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Order ID returned from POST /orders
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    # -----------------------------------------------------------------------
    # Requests
    # -----------------------------------------------------------------------
    PlaceOrderRequest:
      type: object
      required: [user_id, side, type, quantity]
      properties:
        user_id:
          type: string
          description: Trader identifier
          example: alice
        side:
          type: string
          enum: [buy, sell]
          description: Order direction
        type:
          type: string
          enum: [limit, market, ioc]
          description: |
            - **limit**: match at or better than `price`, rest remainder on book
            - **market**: sweep opposite side at any price, cancel unfilled
              remainder. Returns `ErrMarketNoLiquidity` if book is exhausted.
            - **ioc**: fill what crosses at `price` or better, cancel remainder
              immediately (never rests).
        price:
          type: integer
          minimum: 0
          maximum: 100
          description: |
            Price in cents. Required for `limit` and `ioc` (1–100).
            Must be 0 (or omitted) for `market`.
          example: 64
        quantity:
          type: integer
          minimum: 1
          description: Number of lots to buy or sell
          example: 10

    # -----------------------------------------------------------------------
    # Responses
    # -----------------------------------------------------------------------
    PlaceOrderResponse:
      type: object
      properties:
        order_id:
          type: string
          description: Assigned order identifier
          example: "160052-1"
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
          description: Trades generated by this order (empty if no match)
        remaining_qty:
          type: integer
          description: |
            Quantity that rests on the book (limit) or was cancelled
            (market/IOC). 0 when fully matched.
          example: 4

    Trade:
      type: object
      properties:
        buy_order_id:
          type: string
          description: The buy-side order involved in the trade
          example: "160052-2"
        sell_order_id:
          type: string
          description: The sell-side order involved in the trade
          example: "160052-1"
        price:
          type: integer
          description: Execution price in cents (resting order's price)
          example: 64
        quantity:
          type: integer
          description: Number of lots traded
          example: 10

    CancelResponse:
      type: object
      properties:
        status:
          type: string
          enum: [cancelled]
          example: cancelled

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok]
          example: ok

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "user_id: must be non-empty (invalid user id)"